+++ src/Pacco.Services.Vehicles.Core/Entities/Vehicle.cs
@@ -9,13 +9,13 @@ namespace Pacco.Services.Vehicles.Core.Entities
         public string Brand { get; protected set; }
         public string Model { get; protected set; }
         public string Description { get; protected set; }
-        public ushort PayloadCapacity { get; protected set; }
-        public ushort LoadingCapacity { get; protected set; }
+        public double PayloadCapacity { get; protected set; }
+        public double LoadingCapacity { get; protected set; }
         public decimal PricePerHour { get; protected set; }
         public Variants Variants { get; protected set; }
 
-        public Vehicle(AggregateId id, string brand, string model, string description, ushort payloadCapacity,
-            ushort loadingCapacity, decimal pricePerHour)
+        public Vehicle(AggregateId id, string brand, string model, string description, double payloadCapacity,
+            double loadingCapacity, decimal pricePerHour)
         {
             Id = id;
             Brand = brand;
@@ -27,8 +27,8 @@ namespace Pacco.Services.Vehicles.Core.Entities
             AddVariants(Variants.Standard);
         }
         
-        public Vehicle(AggregateId id, string brand, string model, string description, ushort payloadCapacity,
-            ushort loadingCapacity, decimal pricePerHour, params Variants[] variants) 
+        public Vehicle(AggregateId id, string brand, string model, string description, double payloadCapacity,
+            double loadingCapacity, decimal pricePerHour, params Variants[] variants) 
             : this(id, brand, model, description, payloadCapacity, loadingCapacity, pricePerHour)
         {
             AddVariants(variants);
