+++ src/Pacco.Services.Vehicles.Infrastructure/Mongo/Documents/VehicleDocument.cs
@@ -10,8 +10,8 @@ namespace Pacco.Services.Vehicles.Infrastructure.Mongo.Documents
         public string Brand { get; set; }
         public string Model { get; set; }
         public string Description { get; set; }
-        public ushort PayloadCapacity { get; set; }
-        public ushort LoadingCapacity { get; set; }
+        public double PayloadCapacity { get; set; }
+        public double LoadingCapacity { get; set; }
         public decimal PricePerHour { get; set; }
         public Variants Variants { get; set; }
     }
